Refactor Controller.java
Although a moderate size of changes were made in this class, I want to
gather all of them in a single commit as they more or less serve the
same purpose. Now Controller.java contains a field of type
OptimizationProblem to hold the current problem instance and this allows
us to load, export, edit the problem instance on the fly and solve the
current problem instance separately.